サインインありなしの使い分け（ファイルの引っ越し）
