+++ src/app/(authenticated)/MemoEditor/page.tsx
@@ -1,0 +1,5 @@
+import { Tiptap } from "@/component/TipTap";
+
+export default async function MemoEditorPage() {
+  return <Tiptap />;
+}
