+++ src/app/(authenticated)/MemoList/page.tsx
@@ -1,11 +1,9 @@
 // pages/index.tsx
 
 import Link from "next/link";
-import { redirect } from "next/navigation";
 
 import { memos } from "@/pages/api/memos";
 import { Memo } from "@/types/memo";
-import { createClient } from "@/utils/supabase/server";
 
 const PostList = ({ memos }: { memos: Memo[] }) => {
   if (memos) {
@@ -24,15 +22,6 @@ const PostList = ({ memos }: { memos: Memo[] }) => {
 };
 
 export default async function Home() {
-  const supabase = createClient();
-
-  const {
-    data: { user },
-  } = await supabase.auth.getUser();
-
-  if (!user) {
-    return redirect("/auth/login");
-  }
   return (
     <div>
       <h1>無限スクロールのデモ</h1>
