+++ src/app/(authenticated)/MemoViewer/page.tsx
@@ -1,0 +1,3 @@
+export default async function MemoViewerPage() {
+  return <div className="text-gray">MemoViewerPage</div>;
+}
