+++ src/app/(authenticated)/setting/page.tsx
@@ -1,0 +1,3 @@
+export default async function SettingPage() {
+  return <div className="text-gray">SettingPage</div>;
+}
