+++ src/app/(authenticated)/setting/privacy/page.tsx
@@ -1,0 +1,5 @@
+
+export default async function PrivacyPage() {
+
+  return <div className="text-gray">PrivacyPage</div>;
+}
